benchmark: move http_simple.js to http directory
PR-URL: https://github.com/nodejs/node/pull/7094
